lib/accel: fallback when no devices during module_init()
Related to #3266

During discussion in issue above, it came up that if
dpdk_compressdev and dpdk_cryptodev detects no devices
they are still assigned to the relevant accel ops.

Rather than explicitly failing just due to lack of
the devices, each accel module can report -ENODEV.
In such case the module will not be used, but that
error is not propagated to accel framework init.

As a result accel ops assignment will not use such module,
and fallback to available modules.

Reported-by: liuqinfei <lucas.liuqinfei@huawei.com>
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: Ib52dfcb6dc6a1a93b8493b98ffb028f9d9419d6d
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22026
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>