nvme/nvda_tcp: Fix sock buf pending issue
When list of pending recv sockets is processed, socket may be put back
to the list due to lack of sock bufs (and potentially other
reasons). To handle this case socket should be removed from the list
before processing events. To avoid looping on the same socket if it is
readded to the list, we save the last socket and stop when it is
reached. New sockets are always added to the end of this list.

Signed-off-by: Evgeniy Kochetov <evgeniik@nvidia.com>