rpc: Vary JSON-RPC server poll rate based on activity
Since the JSON-RPC server first sends, then receives, if we wait 4ms
between polls then the fastest any RPC can finish is 4ms.

We could have reversed the order of send and recv, but that would have
only fixed the issue for RPCs that complete immediately. Instead, we
take a time-based approach to busy poll for a short window after any RPC
activity.

Change-Id: Ie24933bdd84843bc44d20190ea6cf32c002bb56c
Signed-off-by: Ben Walker <ben@nvidia.com>