[READ ONLY BDEV] new spdk_bdev field - write_disabled
In this commit, we added a new field to the spdk_bdev
structure. The write_disabled field.

The new field indicates if write IO types are disabled
for this bdev.

The following changes were also added:
1. The RPC bdev_get_bdevs was modified to check the new field before printing
   if a bdev supports IO types from write type.

2. The RPC bdev_set_ro was added, it changes the write_disabled
   field to true, preventing a bdev from support IO types from type write.
   It gets a bdev's name, and returns nothing on success (error if failed).

3. The RPC bdev_set_rw was added, it changes the write_disabled
   field to false, making a bdev support IO types from type write.
   It gets a bdev's name, and returns nothing on success (error if failed).

When a bdev's supported IO types get changed, a new event would be sent to it's
open descriptors threads, the event SPDK_BDEV_EVENT_IO_TYPES_CHANGED was added to
signal a user that the bdevs capabilites were changed.

Setting a bdev RO or RW releases any existing claim.
Claiming a bdev while its in a read-only state fails the claim if
the claim promotes a descriptor to read/write state.

Signed-off-by: Ori Evron <oevron@nvdia.com>