nvme: check pthread_mutex_destroy() return value
We have indications that the ctrlr_lock is being destroyed while
held in issue #3401 failures, so check the return value and assert
with an ERRLOG if it fails.

Note that it seems like even when the mutex is held,
pthread_mutex_destroy() doesn't always return failure, but it's still
better to check this return value than ignore it.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ic7091c2a181f953fa447d3eeccbc4350ad5c9698
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23730
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>