lib/bdev: Reset split flag when bounce buffer is used
When read/write IO is sbumitted, we check of IO needs
to be split. That might be true if iovcnt exceeds bdev's
max_sge. However, if memory domain and/or accel sequence
are used, we use a staging buffer and set iovcnt to 1.
In that case no split might be longer required but since
the split flag is still set, we may attempt to split IO
which leads to some weird results or hit an assert in
debug build.
To fix it, reset split flag once staging buffer is
allocated

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: I99a92c35d1791e5c5aa28d29f1692ff7041cc532