lib/jsonrpc: Return busy poller state if there are oustanding requests
RPC poller switches to active polling mode when something is received
or sent on RPC socket. Active polling dureation is 4ms. After that RPC
poller switches back to timed ppoller mode with 4ms interval. If RPC
is asynchronous and its execution time is longer than 4ms, the
response will be delayed because it will be sent out only on the next
RPC poller cycle.

When we have outstanding requests we expect to have response
soon. Reporting busy from jsonrpc server in this case helps to keep
RPC poller in active polling mode and avoid response delay.

Signed-off-by: Evgeniy Kochetov <evgeniik@nvidia.com>