bdev: Bug Fix: QoS group limit didn't work if bdev is opened after group limit is set
For SNAP use case, due to bdev_ready, bdev is opened and I/O channels
are acquired before configuring QoS group limit. However, when QoS is
used with SPDK bdevperf tool, QoS group was configured before bdev is
opened. In this case, group limit was not configured correctly. Fix the
bug in this patch.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>