init: Add framework_[enable|disable]_subsystem RPC
Allow the user to enable and disable subsystems at start up time. This
prevents them from initializing entirely.

A subsystem may only be disabled if no enabled subsystems depend on it.
A subsystem may only be enabled if it does not depend on any disabled
subsystems. The state of the subsystem will be persisted when dumping
the current config to JSON.

Change-Id: I5328d5e20114af11929de933d44de9f8ed509622
Signed-off-by: Ben Walker <ben@nvidia.com>