lib/fuse_dispatcher: FUSE_READDIR handling fixed
The forget param added to spdk_fsdev_readdir's per-fobject callback so the
callback can now decidwe whether to forget the reported fobject.

The fuse_dispatcher's FUSE_READDIR makes use of this param and set it to
true in order to instruct the fsdev to forget the fobject and keep the
reference counter intact as required by the FUSE_READDIR spec.

Change-Id: Ib2713a5690ffb3d59e563e817467196fd277cbec
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>