module/fsdev/aio: fobject/fhandle refcnt protection
FUSE has an issue with commands ordering. For example, in a case where a
user/app/host sends forget for an fobject before a previous IO dealing with
this fobject has been completed.

Consider the following command sequence:
  fobject = FUSE_LOOKUP("my_file")
  FUSE_STATFS(fobject)
  FUSE_FORGET(fobject, 1)

Now, if the app sends FUSE_FORGET before the FUSE_STATFS has been completed
(for example, assuming that the underlying Virtio queue serializes the requests
or due to some bug), and these IOs are executed on different cores, we can find
ourselves in situation of a race where forget deletes the fobject while the
fstats uses it.

Leaving a broader question of what functionality is expected from a driver
in such case, we should protect ourselves in such situation and make sure
that we at least do not crash.

This is exactly what this patch does.

Change-Id: I38d9973e6f3ad7aacc6daa1a388b87a892a35bd3
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>