test: Fix build warnings
There is a lot of 'maybe-uninitialzed' warnings in unit tests. Some of
them are false positives that require changes in real data path code
to workaround. The simple solution is to suppress these warnings for
unit tests. If there is a real issue it should still be reported for
real code.

Signed-off-by: Evgeniy Kochetov <evgeniik@nvidia.com>